rename together_ai to togetherai
