+++ .github/labeler.yml
@@ -174,10 +174,10 @@ integration:stackit:
       - any-glob-to-any-file: "integrations/stackit/**/*"
       - any-glob-to-any-file: ".github/workflows/stackit.yml"
 
-integration:together-ai:
+integration:togetherai:
   - changed-files:
-      - any-glob-to-any-file: "integrations/together_ai/**/*"
-      - any-glob-to-any-file: ".github/workflows/together_ai.yml"
+      - any-glob-to-any-file: "integrations/togetherai/**/*"
+      - any-glob-to-any-file: ".github/workflows/togetherai.yml"
 
 integration:unstructured-fileconverter:
   - changed-files:
