+++ integrations/togetherai/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [integrations/together_ai-v1.0.0] - 2025-10-14
+## [integrations/togetherai-v1.0.0] - 2025-10-14
 
 ### 🚀 Features
