+++ integrations/togetherai/README.md
@@ -3,8 +3,8 @@
 [![PyPI - Version](https://img.shields.io/pypi/v/togetherai-haystack.svg)](https://pypi.org/project/togetherai-haystack)
 [![PyPI - Python Version](https://img.shields.io/pypi/pyversions/togetherai-haystack.svg)](https://pypi.org/project/togetherai-haystack)
 
-- [Integration page](https://haystack.deepset.ai/integrations/together_ai)
-- [Changelog](https://github.com/deepset-ai/haystack-core-integrations/blob/main/integrations/together_ai/CHANGELOG.md)
+- [Integration page](https://haystack.deepset.ai/integrations/togetherai)
+- [Changelog](https://github.com/deepset-ai/haystack-core-integrations/blob/main/integrations/togetherai/CHANGELOG.md)
 
 ---
