+++ integrations/togetherai/pyproject.toml
@@ -27,20 +27,20 @@ classifiers = [
 dependencies = ["haystack-ai>=2.19.0"]
 
 [project.urls]
-Documentation = "https://github.com/deepset-ai/haystack-core-integrations/tree/main/integrations/together_ai#readme"
+Documentation = "https://github.com/deepset-ai/haystack-core-integrations/tree/main/integrations/togetherai#readme"
 Issues = "https://github.com/deepset-ai/haystack-core-integrations/issues"
-Source = "https://github.com/deepset-ai/haystack-core-integrations/tree/main/integrations/together_ai"
+Source = "https://github.com/deepset-ai/haystack-core-integrations/tree/main/integrations/togetherai"
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/haystack_integrations"]
 
 [tool.hatch.version]
 source = "vcs"
-tag-pattern = 'integrations\/together_ai-v(?P<version>.*)'
+tag-pattern = 'integrations\/togetherai-v(?P<version>.*)'
 
 [tool.hatch.version.raw-options]
 root = "../.."
-git_describe_command = 'git describe --tags --match="integrations/together_ai-v[0-9]*"'
+git_describe_command = 'git describe --tags --match="integrations/togetherai-v[0-9]*"'
 
 [tool.hatch.envs.default]
 installer = "uv"
@@ -68,7 +68,7 @@ integration = 'pytest -m "integration" {args:tests}'
 all = 'pytest {args:tests}'
 cov-retry = 'all --cov=haystack_integrations --reruns 3 --reruns-delay 30 -x'
 
-types = """mypy -p haystack_integrations.components.generators.together_ai {args}"""
+types = """mypy -p haystack_integrations.components.generators.togetherai {args}"""
 
 [tool.mypy]
 install_types = true
